Two changes.
One, the imdb_queue change, is targeted towards r1169's fix. It cleans up the possibility that the quality could be not logged, but you have the needed quality set to ANY. It also handles the possibility (an existing issue) that quality is not known, and not logged, but you specify you want to get unknown quality specifically.

The second is pretty major, though the code isn't radical. It's a change to make pirate bay url rewriter perform a basic heuristic analysis on links instead of pure string matching (albiet in lowercase). There's really a lot of power to be had here, but I wanted to try it out on tpb first, since that's the one I have the most trouble with. Tested it out with a handful of streams, and it handled every one with 0 false positives. Huge improvement, as it went from a ~30% failure rate for me, to a 0% failure rate. Lemme know how it does for you, there's plenty of tuning we can do.




git-svn-id: http://svn.flexget.com/trunk@1171 3942dd89-8c5d-46d7-aeed-044bccf3e60c