download subtitles with periscope
An attempt to a subtitles downloader plugin. I'm not sure about the
correct way to do it in Flexget, so I've used the "start" phase and
since it doesn't produce nor consume any entry Flexget reports warnings
about it. Perhaps it should be a output plugin, processing file entries
produced by listdir or find.