Changed test title in test_rottentomatoes
Dropped 'in' from 'Star Wars: Episode I - The Phantom Menace (in 3D)'
as parser was not able to strip this, causing rotten tomatoes to be
unable to find the movie. Test therefore now passes.

Maybe a relevant thing for parser testing, but the aim here is test
rt functionality rather than MovieParser so removed.