Ability to rename all files and rearrange directory
With these changes, all files within a torrent (not just the “main”
file) can be renamed. This is mostly useful for season packs. Series
parsing has been included if the needed fields are not already in the
entry. The containing folder for a torrent (if present) can also be
renamed using a template that renders entry fields.