url_rewrite_torrent411  -> Discover+RSS with login
Sorry for double post.. I had to remove a line.. so here it is.. I had time to test it a bit and seems to work nicely!
It will login only once a day. It will store the login info to the database for 1day
it will search in the html.
download files with login info..
all categories, sub-catogories are supported.
the rss is supported with login info.. works nicely.
formated with pep8 -> only some lines are too longs.

Thks!