Make auto_escape also work with spaces and single quotes
I'm not sure if it might have already worked for Windows, so I didn't change the current behavior for that platform, but in Unix-based systems the option didn't work at all. Spaces were not escaped and single quotes were not escaped. There might be even more things that need escaping, but so far I've only ever come across spaces and single quotes. So far I've fixed it myself by jinja-replacing them ( ```{{ location|replace("'", "\\'")|replace(" ", "\\ ") }}``` ), and now that I haven't encountered any troubles with that in a while I figured it should be a worthwhile fix in the actual codebase.