Bug fixes, season container
New option container_multiple, when set to True, will only add a
container if there are multiple files within the torrent (respecting
main_file_only). For example, this can be used in conjunction with
keep_container set to False to remove the existing container directory
and create a season folder structure for TV shows.