Clean up logging and add verification.
Now use torrent_info_hash to verify that a torrent was added.
I am not sure if this key is available in all cases, so it is
currently in a conditional. Once more testing has been done we
may be able to remove the check.

Also change log.info() calls to log.debug() for obvious reasons.

Fixed a bug loading torrents by adding an empty first param.
I am unsure why this is necessary.