Continued refactoring and expand features
Add method to get torrents from client (used for input). This has a
default list of fields set as a class attribute, which should be set
to whatever we need for the Input plugin.

Move on_task_start to Base class as this is used in all plugins,
including output, input, and cleanup.

Added on_task_exit and on_task_abort to output plugin per @stevezau's
recommendation. This was copied from Transmission plugin.

Began implementing input plugin. Currently does not work and so
it is commented out in register_plugin.