fixed display issue with "flexget plugins"
Some plugins names are too long and don't display correctly when running "flexget plugins". I've added a pipe as a visual split the columns and made the spacing between name and roles columns slightly wider.

The pipe also allow me to pipe the output through awk, which will be useful in the bash_completion script I'm writing for flexget.