Fix race condition with usage of TimedDict in various parts of the code.
Since a TimedDict will expire entries after a certain amount of time,
it's important to be atomic with the way that things are looked up
from a TimedDict.  In particular this access pattern is troublesome:

```python
if key in cache:
  value = cache[key]
```

This pattern is particularly troublesome because the value could become
evicted between the time it was observed to be in the cache, and the
time that it's retrieved from the cache.  When this happens a KeyError
is thrown and this is especially problematic because the code is
almost always written to not assume a KeyError can happen -- because
the key was just proven to be present in the cache.