Convert string-like values to plain strings in Entry
BeautifulSoup, as used in some search plugins, returns strings from HTML
document structures in "enriched" form - instances of unicode/str mixed
in with another class. Those strings carry references to the whole
document tree, which can cause issues when attempting to serialize entry
values for saving to the database.

To avoid that, coerce any values whose type is a subtype of text_type,
but not *exactly* text_type, to text_type, with text_type being unicode
in Py2 and str in Py3. That should ensure we throw away any extra
baggage that might be attached but keep the exact same content.