Add extra fields in HTML input to allow filtering
Currently the entries returned by the HTML input plugin have basically
the URL field. Some more data could be useful to allow filtering in the
filter phase in flexget. Here we add the name of the parent and
grandparent of the link, as well as the classes of the link, the parent
and the grandparent nodes.

With this, suppose the links you are interested in appear in <h2>
headers. You could then use the 'if' filter as

```
test-task:
  html: http://example.com
  if:
    - "html_parent == 'h2'": accept
  <some_output_plugin>
```