new search plugin for private tracker torrentday
i used torrentleech as a starting point. The main difference being that TL uses uname/pass to login, generate the cookies & access the search pages. TD's login page has captcha, so instead i put the 3 cookies it needs as required keys. this should be fine, because in my browser they all have expiry date of 2038. Beyond that, the 2 sites have a slightly different CSS so i'm look for some different classes and divs (this is my first PR ever, so plz be gentle)