[add] btn - exact_only option to omit second search
If the title to be searched ends in a parenthesis, and there are no
results on the first search, btn automatically runs a second search
omitting the text in parenthesis. This can cause unwanted results to be
accepted. This change adds a second config type, a dict with api_key
and exact_only. When exact_only is set to yes, the second search will
never be run.