[fix] next_series_seasons - Limit size of emit in one run
Given a season identifier of S2014 and backfill enabled, it is possible
for `next_series_seasons` to emit all the way down to zero. This change
prevents it from emitting if `begin` is not set and the group of
seasons to review is more than 15, or if `begin` is set and the
difference between the latest season to review and `begin` is more than
30. In both cases the user will be required to set or adjust `begin`
before any emit will occur.