[fix] next_series_seasons - Use latest download for rerun entry generation (#1829)
* [fix] next_series_seasons - Use latest download for rerun entry generation

Given begin episode of S2014E01, S2017 downloaded in history, and
backfill enabled, `next_series_seasons` will emit S2014, S2015, S2016,
and S2018. If those are accepted and it’s allowed to rerun, it will
then look for S2015, S2016, S2018, and S2019, because when each of the
previous seasons was accepted, a rerun entry was created for the season
immediately after the accepted season regardless of whether that “next”
season was also accepted in that run. This change uses the latest
download to determine the next season to search.

* Use any instead of list comprehension