[add] sort_by - Option to ignore leading articles
Adds an additional option, `ignore_articles`. When set to yes, the sort
will ignore the English articles ‘the’, ‘a’, and ‘an’ when they appear
as the first word in the field to be sorted. Alternatively, you can
also specify a custom regular expression to be used, for use with other
languages or to otherwise customize the articles that are ignored.