[fix] updated the nyaa plugin to use the new url. (#1887)
* udated the nyaa plugin to use the new url.

nyaatorrents.se voluntarily took itself offline because of some rumors that the "police" were coming after them. a few alternate versions popped up shortly after, and this seems to be the defacto one. I simply took the old nyaa plugin and updated the urls and values wherever necessary.

* updated the nyaa test

* found a torrent that actually exists

* found another possible url that needed to be updated.

i'm unfamiliar with testing to this extent, so I'm trying my best to figure it out.