[add] sonarr_list - More options
* Add couple more options.

~ Added option to set quality profile ID (defaults to 1/any) -- You can find quality profile/profile ID at http://localhost:8989/api/profile?apikey={APIKEY}
~ Added option to set season folder -- True/False Default is False
~ Added option to set monitored status of whole series -- True/False Default is True

*** NOTE
One thing I don't understand is the why there is a profileId and a qualityProfileId as it seems only the profileId actually does anything, but I'm leaving it in as it seems it's part of the API call anyways versus profileId not being any part of the Sonarr API..

* Make option for rootFolderPath

~ The way I had to set this up was that it uses the ID instead of a string path and it would take the number from "rootFolderPath" config and subtract it by 1, so 1 = 0, 2 = 1, so on and so forth. You can find the rootFolderID here http://localhost:8989/api/rootfolder?apikey={APIKEY}

*** Note
Again I don't really know if this messes with anything for trakt/sonarr to only flexget. Doesn't mess anything up for trakt to sonarr though.

* Change from camel case to snake case

* Add seriesType option

~ Adds Option for seriesType, string value of standard, daily, or anime. (default value is standard)

## Currently working on getting an option to set tags, but it's being very tricky..

* Requested changes

~ Changed root_folder_path to be an int instead of a number, did the same to profile_id as it's supposed to be an int anyways as well..
~ Added an enum to series_type as it's a closed list

* Add option to set tags

Allows you to set tags | Has to be by ID and has to be set up like example [1,2,3,4,5] (replacing any numbers with your tag IDs) Tag ID's can be found at http://localhost:8989/api/tag?apikey=API | Requires you to have pre-setup any tags you plan on using. | Default value is [0] (not passing any tags)

* comment on how rootFolderPath is setup.

* Add something to comment.

* One more addition to comments.

* Make comments less than 120 chars & multi line.

I hope this is better >..<

* change a word in one of my comments.

* Fix codacy complaints