Add metacritic score parsing
Add metacritic score parsing from IMDB movie pages. 

Note: I was not able to test this. I think this should work, but the HTML class is defined as follows in IMDB:
<div class="metacriticScore score_favorable titleReviewBarSubItem">

I'm not sure if the class statement on line 325 needs to include all three classes to make the match.