make_rss: default encoding to UTF-8
In Python 3 all strings are in UTF-8, and the RSS XML generator does not
transcode. So, if we let the RSS XML be generated with the default
encoding that is not UTF-8 from entries that use default strings
(UTF-8), then the generated XML will be broken, and browsers that parse
it will report invalid characters.