Plugin `make_rss`:  add and document options, emit enclosures, default encoding to UTF-8 (#2324)
* make_rss: make timestamping optional

Useful when need the generated file to be exactly the same if the
entries did not change.

* make_rss: document item title option
* make_rss: rsstitle and rssdesc config options
* make_rss: take pubdate from entry if specified

It will be specified if the entry comes from the input RSS plugin.

* make_rss: output enclosures
* make_rss: default encoding to UTF-8

In Python 3 all strings are in UTF-8, and the RSS XML generator does not
transcode. So, if we let the RSS XML be generated with the default
encoding that is not UTF-8 from entries that use default strings
(UTF-8), then the generated XML will be broken, and browsers that parse
it will report invalid characters.