Prevent accidental downloading of magnet links
On high speed connections setting paused to `False` might cause the
magnet file to be downloaded event though `add_paused` was defined.
This issue is noticeable especially when transmissionrpc times out.

Changing the paused status (Flexget#462) is meant to address an issue
with handling `content_filename` with magnet links but there is no
need to change the status if `magnetization_timeout` is zero.

Original issue:
https://web.archive.org/web/20150515004451/http://flexget.com/ticket/2911