Sort series entries based on quality
To get the best quality in a given timeframe or within a single task
execution the entries should be sorted by quality. If the entries are
not sorted, then the first seen entry is wrongly assumed to be the
best quality entry, which messes up both single task execution and
multiple executions when using timeframe option.