[add] from_transmission: Add 'transmission_errorString'
This adds the field `transmission_errorString` when parsing torrents with `from_transmission`. This is useful for taking action depending on the specific type of error reported by the tracker, rather than just the presence of an error. For example, you may want to clean up torrents that have been removed by the tracker, but leave torrents when the tracker is temporarily down.