Handle both old and new case for database_uri
I found this case when installing the latest version of flexget. 
It defaults to an empty string instead of, I presume, the previous default of `None`.

I'm not sure if it's worthwhile keeping the `None` as a case but figured it was the 
"safer" option just in case. Also feels like `if not self.database_uri` should be just enough as well?