[fix] iptorrents: parse new search page redesign
A lot of the css classes were removed on the IPT site, which the
previous code relied on. Instead, this parses the table headers.

The html.parser doesn't parse tables correctly so I switched it to
html5lib. Specifically, '<td><td><td>' would create a nested
structure instead of siblings.