[fix] Pyload: with packaging installed and/or with using pyload-ng
Pyload is currently not working with packaging installed (on both stable and ng)
and it is probably not working on pyload-ng with or without packaging installed.

What is wrong
=============

For stable
-----------

GET method for /getServerVersion will not work as we have to be logged in. This
returned completely wrong data which were later acted on.

For pyload-ng
-------------

This was working only when packaging was installed. In #2642 there was attempt
to work without packaging, but IMHO it simply turns the version checking off and
pyload is always considered non ng.

How to fix it
=============

I chose rather simple approach. We currently don't really need exact pyload
version. We only need to know if we are on stable or ng. And we can easily
detect it by testing login response, no need to do whole magic around.