[add] torznab: More complete RSS XML entry data
I needed access to the `size` child of the RSS `<item>` element like the `rss`
input plugin provides.  While trying to address that I noticed what seemed
like an unnecessarily limited and hard-coded approach to extracting the
`<item>` elements children, so this change implements a more generalized
approach.

I would like to use a more generalized approach for the `<torznab:attr>`
children as well.  The current implementation is, however, sufficient for now
and ripping that out would be more invasive.

In fact, I'd love to see FlexGet use something like
[`xmltodict3`](https://github.com/dart-neitro/xmltodict3) wherever data is
being extracted from XML and inserted into entries so that using RSS or other
XML in FlexGet is consistent throughout.