[change] html plugin add stop_when_404 increment option (#3240)
Some sites will return 404 at pagination stop. When 'stop_when_404'
is set to 'false', the HTTPError thrown by requests will bubble up
(default and existing behaviour) - aborting the task. If it is set
to 'true', the HTTPError will be caught and iteration will halt as
if an empty page was returned and 'stop_when_empty' was set to 'true'.