Add suite of test cases to the sftp_{list, download, upload} plugins.
This change uses a local test sftp server to assert the behaviour of the
various sftp plugins. By using an actually test server it opens the
door to moving to Paramiko if needed (see
https://github.com/Flexget/Flexget/issues/3619) while ensuring that
the current functionality doesn't break.

Note: In addition it also enables a host hey to be specified via the
plugin config for the list and upload plugins. This makes it possible
to use the test sftp server without adding an entry to known_hosts.