If a sftp_download entry represents a symlink and delete_origin is true, only remove the symlink.
Currently is a entry represents a symlink, the target of the symlink gets removed rather than the symlink itself. This change corrects that and ensure that in the case of a symlink entry, only the symlink itself get's removed.

This also contains a fix for the test_sftp_server to use absolute() rather than resolve() when canonizing paths. Without this change lstat (which shouldn't follow symlinks which resolve does) fails to work correctly.