Don't override default timezone in the scheduler
Since the change in #3453, the code was passing a non-pytz timezone,
causing #3603.

All we want to do is to tell appscheduler to use the local timezone, and
as seen in the documentation for [appscheduler.schedulers.base][1],
the local timezone is already the default value, so there is no need
to pass this option.

[1]: https://apscheduler.readthedocs.io/en/3.x/modules/schedulers/base.html#module-apscheduler.schedulers.base