Add more detailed assertions for entries generated by the sftp_list plugin.
Previously the test cases only focused on if a certain entry existed, not that it contined the correct data. As such this change adds assert_entries() method which allows more detail about the entries generated to be specified as well as checking if any erroneous entires were also generated. This change preceeds the change of symlink handling in the sftp_list plugin to make it easier to assert the URL change in the test.

This change also includes a small bit of cleanup in terms of the task names to make it clearer which variables are been set in that test case.