If a sftp_download entry points to a symlink and delete_origin is true, only remove the symlink, and not it's target.
This changes the behaviour when dealing with entries representing
symlinks. If entry previously represented a symlink it would remove it's
target or in the case of a directory it's contents. This changes that
behaviour such as in the case of a symlink, only the link will be
removed, but not it's target.