Fix letterboxd.py movie data retrieving
Currently letterboxd in their watchlist the data-film-slug is just the end part of the URL to the film.

As a quick example for the movie "/film/asteroid-city/" the `data-film-slug` will be populated with "asteroid-city". 
Without this fix the generated path would be `http://letterboxd.comasteroid-city` causing failure due to DNS resolve failing.