refactor how HTTP requests and authentication is handled
Previous code was a bit scattered, with the authentication being methods
on the plugin and the requests being handled in the OmbiEntry class. I
have created a new class called OmbiRequest to handle both. I based this
design on TVDBRequest which seemed to be a decent implementation. The
only complaint I have is how much I have to pass around OmbiRequest, it
might be better to handle it as a module level variable, but I'm not
sure if that is appropriate.