add requestId propery to OmbiEntry class
If you use the /api/v1/Request/movie endpoint it will return all the
movies that have been requested. For some reason this endpoint does not
return the requestId. Then is needed to perform many other API actions,
like changing the status of the request to approved or denied.

I have added a property to the OmbiEntry for requestId. This property
will use the api/v2/Search endpoint which does return the requestId.