fix issue with how Ombi items are filtered based on status
While testing I found out that an item in Ombi can be both approved and
denied at the same time. This meant I had to rework the filtering logic
to account for this.

I also found out that the /Search and /Request API's both return "items"
but the fields are different? For example when you perform a movie
search it will return a movie where the ID is the id for the movie and
the status is the status of the movie. But when you list all movie
requests you get a list of requests where the ID is the request ID and
the status is the status of the request. This seams obvious now but at
the time it wasn't because the structure of the JSON response for both
are nearly identical.