add ability to request TV shows, seasons and episodes in Ombi
I'm still recovering from a cold so I don't have the energy to break
this up into multiple commits.  The main thing I did refactor how the
OmbiTv class works. It no longer accepts multiple Id's as I can only get
the tmdb_id to work with the Ombi APIs. I refactored how the
make_request method works. This way it was easy to implement it for both
OmbiMovie and OmbiTv. I also added a field called "ombi_title" so that
it prints the full title depending on if you have a show, season or
episode.