add ombi_list plugin (#3481)
[add] ombi_list plugin

* Ombi input plugin

* Corrected the way the requests plugin was imported to use the flexget version for better error handling

* Removed unused json import

* Added error checking for missing IMDB ID when generating a URL.

Probably a should be a function but wanted to commit this quick fix as it has caused me issues and was only in my local version.

* move ombi input plugin to managed lists

* format and fix imports

* fix import deprication warning

* add the list interface to ombi

* implement suggestions from PR review #2109

* implement remaining list interfaces

* implement feedback from pr

* refactor movie/tv logic into parent class

* add doc string and better type hints for auth func

* black formatting updates

* Revert "refactor movie/tv logic into parent class"

This reverts commit 2e884d902207a9626671bc913d79a0ba17e8ed62.

* change static method to class method

* inverse the if statement

* use url instead of base_url and port

* some linting fixes

* revert movies back to using IMDB

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* refactor how HTTP requests and authentication is handled

Previous code was a bit scattered, with the authentication being methods
on the plugin and the requests being handled in the OmbiEntry class. I
have created a new class called OmbiRequest to handle both. I based this
design on TVDBRequest which seemed to be a decent implementation. The
only complaint I have is how much I have to pass around OmbiRequest, it
might be better to handle it as a module level variable, but I'm not
sure if that is appropriate.

* add error handling when making calls to the Ombi API

In most cases we will log an error message and return None, which
should allow the processing to continue.

* add custom exeception class for the Ombi API

Sometimes the Ombi API will return a 2xx status code but the response
still contains an error message. This custom exception class will
handle those cases.

* represent plugin configuration as a typed dict

* refactor how items are retried from ombi and how the cache works

* allowing the adding of items with a configurable status

This commit adds a number of new methods to the OmbiEntry class to mark
an entry as approved, available, requested, or denied. It also refactors
the add() method to support adding an entry with a status.

* fix the return value for already_requested when item not requested

* allow ombi_list to work with movies using imdb_id or tmdb_id

* add annotations backports to allow use of newer type hinting features

* add ability to specify how items are removed

This commit adds the on_remove config field to the ombi_list plugin.
This allows users to specify what happens to items when they are removed.
The default is to `delete` which clears the requested status from the
item in Ombi.

* use all the supported ids when searching the list for entrys

* change the API endpoint back to v1

* fix error checking with result is a list and not a dict

* add requestId propery to OmbiEntry class

If you use the /api/v1/Request/movie endpoint it will return all the
movies that have been requested. For some reason this endpoint does not
return the requestId. Then is needed to perform many other API actions,
like changing the status of the request to approved or denied.

I have added a property to the OmbiEntry for requestId. This property
will use the api/v2/Search endpoint which does return the requestId.

* fix issue with how Ombi items are filtered based on status

While testing I found out that an item in Ombi can be both approved and
denied at the same time. This meant I had to rework the filtering logic
to account for this.

I also found out that the /Search and /Request API's both return "items"
but the fields are different? For example when you perform a movie
search it will return a movie where the ID is the id for the movie and
the status is the status of the movie. But when you list all movie
requests you get a list of requests where the ID is the request ID and
the status is the status of the request. This seams obvious now but at
the time it wasn't because the structure of the JSON response for both
are nearly identical.

* make avaliable a seperate status than approved/denied

I also refactored movie filtering into its own function.

* handle url for episodes

* add ability to filter episodes by request status and availability

* add ability to request TV shows, seasons and episodes in Ombi

I'm still recovering from a cold so I don't have the energy to break
this up into multiple commits.  The main thing I did refactor how the
OmbiTv class works. It no longer accepts multiple Id's as I can only get
the tmdb_id to work with the Ombi APIs. I refactored how the
make_request method works. This way it was easy to implement it for both
OmbiMovie and OmbiTv. I also added a field called "ombi_title" so that
it prints the full title depending on if you have a show, season or
episode.

* refactor _find_entry to work with TV shows, seasons and episodes

This was needed to properly support list_match.

---------

Co-authored-by: bagobones <bagobones@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>