chore(deps): revert "Remove ruff from dev dependencies (#4454)"
This commit reverts the change from pull request #4454.

Ruff is being re-added to the development dependencies. It was previously removed in #4454, but as @BrutuZ pointed out, this caused the Ruff extension in Visual Studio Code to fall back to its outdated, bundled version of Ruff.

Although this can be addressed by installing Ruff globally, that approach is not ideal because other projects may depend on different versions of Ruff.

Given the widespread use of Visual Studio Code among contributors, this change is being made to ensure a smoother and more consistent development experience.